+++ src/Entities/SuggestedPostInfo.php
@@ -5,13 +5,12 @@ namespace Longman\TelegramBot\Entities;
 /**
  * Class SuggestedPostInfo
  *
- * Contains information about a suggested post.
+ * This object describes a suggested post.
  *
  * @link https://core.telegram.org/bots/api#suggestedpostinfo
  *
- * @method string             getState()    State of the suggested post. Currently, it can be one of “pending”, “approved”, “declined”.
- * @method SuggestedPostPrice getPrice()    Optional. Proposed price of the post. If the field is omitted, then the post is unpaid.
- * @method int                getSendDate() Optional. Proposed send date of the post. If the field is omitted, then the post can be published at any time within 30 days at the sole discretion of the user or administrator who approves it.
+ * @method SuggestedPostPrice      getPrice()   Optional. Price of the suggested post
+ * @method SuggestedPostParameters getParameters() Optional. Parameters for the suggested post
  */
 class SuggestedPostInfo extends Entity
 {
@@ -21,7 +20,8 @@ class SuggestedPostInfo extends Entity
     protected function subEntities(): array
     {
         return [
-            'price' => SuggestedPostPrice::class,
+            'price'      => SuggestedPostPrice::class,
+            'parameters' => SuggestedPostParameters::class,
         ];
     }
 }