+++ src/Entities/UniqueGift.php
@@ -16,9 +16,9 @@ namespace Longman\TelegramBot\Entities;
  * @method UniqueGiftModel    getModel()             Model of the gift
  * @method UniqueGiftSymbol   getSymbol()            Symbol of the gift
  * @method UniqueGiftBackdrop getBackdrop()          Backdrop of the gift
- * @method string             getGiftId()            Identifier of the regular gift from which the gift was upgraded
  * @method bool               getIsPremium()         Optional. True, if the original regular gift was exclusively purchaseable by Telegram Premium subscribers
  * @method bool               getIsFromBlockchain()  Optional. True, if the gift is assigned from the TON blockchain and can't be resold or transferred in Telegram
+ * @method bool               getIsBurned()          Optional. True, if the unique gift was burned and isn't available anymore
  * @method UniqueGiftColors   getColors()            Optional. The color scheme that can be used by the gift's owner for the chat's name, replies to messages and link previews; for business account gifts and gifts that are currently on sale only
  * @method Chat               getPublisherChat()     Optional. Information about the chat that published the gift
  */